♻️🚚: move nictype test into package
